Looking at the Makefiles in the xpackages.
Added Makefile for xcontents and simplified the `localinstall` target.

New `check` target for automated checking of the "test" files. Some
additions and minor changes in the individual Makefiles for this. Not
much achieved though; a lot of these tests are designed to be run
interactively.

Tests currently included are for xfootnote, xhead, xinitials, and xtheorem.

Everything does its best to leave things nice and tidy, but there
might still be some residual files left around. Should probably be
like expl3 and move everything to a temporary directory.