Undiscussed change to remove \engine_
This was starting to bug me; it was pretty clear that there wouldn't
be any other (or many more) functions that could start with \engine_,
so the engine-checking functions are now
  \xetex_if_engine:NTF
  \luatex_if_engine:NTF

This eliminates one "module" name, a task that always gives me great
happiness.

Actually, these could also be
  \mode_if_xetex
  \mode_if_luatex
but I prefer things this way.

Hope this doesn't break any of your code :) (I'd be surprised!)