Created template-new to implement discussion ideas
This new version of template splits key declaration and implementation into two parts. This allows a clearer flow for editing and restricting templates.

The key declarations look like:
  integer    : my-integer = 1,
 +length     : my-length, % No default, global
  function.2 : my-fucntion = (#1:#2),
  code       : arbitrary

Currently, there is no documentation but this will be written over the next few days. There is still discussion to be completed on a number of ideas, and so the code may change before the documentation is written.

The implementation currently uses property lists extensively. If this is a problem for storage, it would be possible to "pack" these inside token lists.

git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@1518 de43f980-851b-0410-b2f7-c40aca1f87e0