Added "hyphenation" template to xgalley.
This implementation is somewhere between "TeX" and "std" in galley2.
The low-level stuff is hidden inside the code part, but there is a bit
more flexiblity than in the original "std". 

At the moment penalties are hard-coded: should this be done using a 
second template with instances? Also, should this key accept an
abitary value via the "unknown" setting?


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@1831 de43f980-851b-0410-b2f7-c40aca1f87e0