Add \fp_to_tl:NN family to l3fp
This is an 'intermediate' function, which takes a few settings to 
determine exactly how it works. Using a suite of related transfer
functions seems confusing in comparison: comments?

The idea here is that you can use l3fp to turn data back into something
to present to the user, with speed gained by knowning the internal
fp structure (cf. siunitx, which has to round without knowning the
number of digits).

Open issues include expected behaviour of infinite/undefined results,
rounding in integer part and possibility of rounding to significant
figures (see again siunitx)