fix source3 compilation after l3doc 'improvement'
so \begin{variable} can now be used in documentation or
implementation sections and its behaviour switches
accordingly.

but if you *want* a variable documentation section inside
the implementation (see l3token, l3xref) there's a command
for that, too: \begin{variabledoc}

not sure if this is the best approach. i'm thinking that
we could do the same thing for \begin{function} being
an alias for \begin{macro} inside the implementation; I
think this is sensible but before it can happen I need to
unify the syntax between the two environments.

but perhaps it is better just to call them different things,
such as functiondoc and functioncode. We don't often have
commands that change behaviour based on their context, do we?