First part of galley template documentation
There are a number of questions outstanding here: the documentation
covers the current approach, but even the number of arguments needed
by each object type may change!

The 'justification' and 'line-breaking' objects still need to be done.
The later is little more than a wrapper around TeX itself, but that
may be hard to document cleanly.