update everyhook translation
can no longer use \toks_ commands, although in
initex mode we'll still need a \toks_new:N
allocator (or handcode each toks register).

also uses the new left/right seq push and pop
functions, which is nice

the design of the package is copied from everyhook,
and Steve Checkoway is happy for us to steal his
work; if this is ever released through expl3 he's
happy to update his package to be a thin wrapper
around this. (last time we spoke, anyway.)