Finalise 'updated' data
Looking back over the SVN log and CTAN release statements, the first CTAN
version of l3kernel (as opposed to expl3) is SVN 2448 dated 2001-06-05. This
seems like a good 'baseline' to take for the information, as working
through all of the 'big bang' will be tedious and error-prone.

Notes on the approach taken:
 - 'added' means the first time a base function is added to l3kernel.
 - Renames count as 'added', as the base function has a new name.
 - If \foo:Nn is added first, and \foo:NV is added later, 'added'
   is used for when \foo:Nn appears and 'updated' for \foo:NV.
 - Functions moved from l3packages/l3experimental/l3trial go down as
   'added' when they first go into l3kernel, not when they first appear
   elsewhere.
 - For items added as 'experimental' and then moved to 'stable', the
   date of the move is 'updated'.
 - Internal changes to functions go down as 'updated' if they are
   significant. This is clear if the semantics alter at all, but
   also applies where functions are re-written for speed, even if the
   behaviour is unchanged (just in case).

Obviously, all new additions/alterations to l3kernel should be logged
appropriately!