Register numbers are not hidden anymore in tlg files.
Instead, a given number of registers is allocated at the end of
regression-test.tex. If a change in some l3kernel module shifts
register numbers (e.g., when adding or removing scratch registers),
we need to change the number of such registers allocated in
regression-test.tex, to keep register numbers stable in tlg files.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@2833 de43f980-851b-0410-b2f7-c40aca1f87e0