Register numbers are not hidden anymore in tlg files.
Instead, a given number of registers is allocated at the end of
regression-test.tex. If a change in some l3kernel module shifts
register numbers (e.g., when adding or removing scratch registers),
we need to change the number of such registers allocated in
regression-test.tex, to keep register numbers stable in tlg files.