Significant improvements to xgalley system
This check-in deals with a number of issues, with a clearer 
definition of the different between local and global galley items.

The code and template levels are separated out to make checking that
documentated interfaces are used easier.

Some work remains to be done: to be fixed over the coming days.