Add a "native-drivers" option to expl3
This loads the (basic) graphics support we have, rather than using the
graphicx and color packages. This is undocumented at present as it
is mainly for internal testing. If this looks useful, this way of
loading expl3 could be used in the test files to avoid issues between
different TeX systems when testing.