Add \tl_display:n, as part of a new module tl-convert.
One aim of the tl-convert module is to convert from an arbitrary
token list to a "safe" token list, detecting special tokens, and
keeping track of category codes and character codes. This aspect
is done.

The second aim is to do a somewhat reverse conversion: given a
string such as "\c{left}\(x+y\c{right}\)", produce the token list
"\left(x+y\right)". This will be used in the replacement text
of regular expressions acting on token lists.