Add 'l3dt' to l3trial
This module introduces the idea of a 'data table', an array-like
structure for storing records with fields.

The data structure is similar to that used in datatool, and is
evenutally intended to enable similar functionality to be built on
top of the code.

The reason for creating this module is to support other experiments in
two areas: meta-data and module management. Both of these require
two-dimensional arrays of data and therefore lend themselves to
working with a structured data type.

The structure here is inspired by both datatool and the approaches
taken in the current l3prop and l3seq modules. The current datatool
structure is basd on some ideas from Morten which are similar to the
way l3prop used to work. l3dt is more like the updated l3prop.

At present, there is no mapping or more advanced selection from
the data table. Some will be added when needed, other parts if/when
it is clear that the entire idea is useful.