Added scan marks to l3quark as an experimental feature.
I'm really not sure where scan marks should go, nor how
to indicate that they are only an experimental feature.
Please feel free to move it elsewhere, but we'll definitely
need it at some point.

(1) They are needed in l3regex in contexts such as

  <instruction 1>
  ...
  <instruction N>
  \s_stop

so that some instruction can skip to the end, but that
TeX doesn't do anything if no function breaks to \s_stop.

(2) They are used in l3galley to "cancel" the previous `\par`

(3) They will be used as the first token of every
floating point number.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3056 de43f980-851b-0410-b2f7-c40aca1f87e0