Overhaul of user commands in l3regex.
Now we've got "\regex_set:Nn", "\regex_gset:Nn",
and "\regex_const:Nn", and the following

  \regex_match:nn(TF)
  \regex_count:nnN
  \regex_extract_once:nnN \regex_extract_once:nnN(TF)
  \regex_extract_all:nnN  \regex_extract_all:nnN(TF)
  \regex_replace_once:nnN \regex_replace_once:nnN(TF)
  \regex_replace_all:nnN  \regex_replace_all:nnN(TF)
  \regex_split:nnN        \regex_split:nnN(TF)

as well as the same with the first argument a precompiled
regular expression. That's 51 user functions! I minimized
the number of auxiliaries used to implement them.

Normally, all of those functions work with token lists,
any remnant of the word "string" in the doc means I haven't
updated that part of the text. However, the support for
matching csnames is not there yet, and nothing at all is
done in the replacement text catcode-wise.

Furthermore, unbalanced braces can be a problem:

  \regex_extract_once:nnN { . } { {ab} } \l_foo_tl

breaks horribly, trying to store "{" in "\l_foo_tl".