Implement \c{...} for the regex (not replacement).
I removed \C and changed the undocumented \c: I went for

   \cC = any Control sequence
   \cB = any Begin-group character
   \cE = any End-group character
   \cM = any Math shift character
   \cT = any alignment Tab character
   \cP = any macro Parameter character
   \cU = any superscript character (Up)
   \cD = any subscript character (Down)
   \cS = any blank Space character
   \cL = any letter
   \cO = any other character
   \cA = any active character
   \c{...} = a control sequence whose csname
           is matched by the regex "..."

This is a little bit more messy at the code level than
using the hexadecimal value of catcodes, it is also more
user-friendly but less old-time-TeXer-friendly (I think).

Also, a more powerful alternative would be to make

  \c<catcode><character>

match the <character> with the given <catcode>, but this
doesn't interact nicely with character classes.