Fix two bugs in the tl rescan functions.
(1) If the rescan function was used in an alignment,
with & in the argument, that & would trigger the formation
of a cell before the end of the rescan function
(because the \scantokens{#3} appeared without surrounding
braces or \group_align_safe_begin:).

(2) If the rescanned token list was a brace group,
outer braces were removed (because of how TeX treats
delimited arguments).



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3066 de43f980-851b-0410-b2f7-c40aca1f87e0