Tidy up the allocation system properly
Bruno points out that the maximum register number changes between
LuaTeX and older engines affect how you do allocations. The maxium values
are

*LuaTeX*
\chardef   1114111
\mathchardef 32767
registers    65535

*XeTeX*
\chardef   1114111
\mathchardef 32767
registers    32767

*pdfTeX*
\chardef       255
\mathchardef 32767
registers    32767

(Currently there is a bug in LuaTeX which means \mathchardef accepts
32768, but we'll ignore this!)

This means that the choice for assigning \int_const:Nn is the same for
LuaTeX and XeTeX (use \chardef), while the choice for register assignment
is the same for pdfTeX and XeTeX (use \mathchardef). The updated code
should tidy up all of this, and also uses \(math)chardef as far as possible
in \int_const:Nn rather than using up registers.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3084 de43f980-851b-0410-b2f7-c40aca1f87e0