Some initial thoughts on metadata
This new code is an exploration of what can be achieved in two areas.
First, there is storing data in a data table (as supported by the l3dt
concept). The code for this supports relational concepts, so that for
example several authors with the same address result in the storage of
only one version of the address. Secondly, there is an exploration of
how a user interface might look to allow multiple authors and multiple
addresses, etc. This is all based on experience with achemso, where both
ideas are required.

This is not attempting to cover all of the ground explored in xfrontm,
which also deals with titles and output. At this stage, I'm imagining
a rather basic LaTeX2e output approach, essentially building \@author
to contain affiliations in some way. This might look something like the
approach in the existing authblk package, as this will work with LaTeX2e
code now.

Longer term, there needs to be either a template-based way of getting 
the information back out, a set of defined 'extract data' functions or
a description of the data structures such that others can use them. The
later seems a bit weak, so I'd favour one of the other two. However,
first I'd like to see feedback on the basics, and perhaps add things
like titles, parsing of names (do we want BibTeX-like name formatting?),
etc.