Work on the expandable parse of xparse
The main aim here is to allow 'D' arguments to correctly grab
input of the form \foo[[bar]]{baz} by implementing the same
method as is used non-expandably. A few extra csnames are needed,
and the approach is therefore a little dense.

The second change is to make internal names a bit clearer, 
to help with debugging.

There is a bug fix when grabbing 't' arguments expandably, which
was not correctly picked up in the test file.

As we have \str_if_eq:nnTF, this can be used rather than
\t_if_head_eq_charcode:nNTF for expandable tests. The later
will match something like \foo{[oops}, while the string
test will not.