Expandable arguments should not use 'l' arguments
When searching for an optional argument in an expandable manner, a FALSE
test means 'recycling' the grabbed token(s). This has to be done with
braces added, as we cannot be sure if what was grabbed had braces before
or not. Thus something like and arg. spec. {ol} will potentially go wrong
when the optional argument is omitted.

As expandable support is limited, it seems most sensible to remove l arguments
here rather than take complex routes to fix this. At the same time, there are
a few simplifications that can be made to the internals.