Further harden tests for matching LaTeX3 code
The approach now is to first use LaTeX2e's mechanism for package
date, so we get a warning early on if the support is too old. Then there
is an explicit bail-out test as soon as possible, not assuming that
even \ProvidesExplPackage is available. This covers the case where
very old expl3 installations are about (see for example LaTeX bug #4178),
at least better than the previous approach.