Revising xparse argument grabbing: expandable functions
Grabbing arguments generates various internal functions. If an error
occurs, for example a runaway argument, the name of the internal
function should give some clue to the end user as to where the problem is.
Thus the internal functions should be named to be 'user readable'. A 
simple scheme is to use" \<function> arg <number>", as this should
not clash with anything else (it contains spaces) and points to the
appropriate argument.

Revision at this stage just changes expandable functions as they already
require a 'full set' of dedicated functions. Revision to non-expandable
functions will require a little more work.

At the same, a couple of internal changes. \q_nil is the canonical marker
to use for retaining braces, rather than "x". Using \cs_to_str:N here avoids
needing to worry about \tex_escapecahr:D, and is safe as the spaces in
internal names remove them from general clashes.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3131 de43f980-851b-0410-b2f7-c40aca1f87e0