Add code to allow implementation of \\
Currently, this is just the same approach as taken by LaTeX2e.
There are some thoughts in ltspace.dtx which suggest that a better
approach may be possible, but there is no detail! One obvious possiblity
is to simply ignore a \\ in vertical mode, rather than issue an error
as happens currently.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3211 de43f980-851b-0410-b2f7-c40aca1f87e0