Add code to allow implementation of \\
Currently, this is just the same approach as taken by LaTeX2e.
There are some thoughts in ltspace.dtx which suggest that a better
approach may be possible, but there is no detail! One obvious possiblity
is to simply ignore a \\ in vertical mode, rather than issue an error
as happens currently.