Guard against active characters in file names
This is a common LaTeX2e issue, and so we should trap it out.
The same will apply for example to graphics (which with LaTeX2e
one needs grffile).

We may still need to address non-native encodings, although
I would favour the 'use a UTF-8 engine' response in those
senarios.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3225 de43f980-851b-0410-b2f7-c40aca1f87e0