Deal properly with the filename issues
The logic here has to cover two competing requirements:
 - File names should expand to allow variables (for example ones including \jobname)
 - Active characters in file names should treated as strings (for example "~" is
   common on Windows, "_" is common generally, etc.)

The best approach is to make the actives safe, then expand, then convert to a string
so that there are no further dangers and so that spaces can be checked for properly.
As this comes up in a few places, an internal function to do the conversion is the
best approach.