Explain what \prg_case_str:xxn really expands.
The commented-out comment is currently irrelevant, but we must add it 
when l3rand comes of age: given the current implementation of 
prg_case_str:xxn, the test string must be equal to itself for the loop 
to end. But that may be broken by random numbers:

   \prg_case_str:xxn { \rand_get: } { } { }

would break in unknown ways (probably try to run a quark, or something 
like that.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3250 de43f980-851b-0410-b2f7-c40aca1f87e0