Remove xo-alloc.dtx
With e-TeX allocating inserts, the code here was largely redundant

The parts still required have been moved to other places, so they
are defined near where they are actually used. This leaves open
a few more refinements, but those will be picked up as the rest
of the code is reviewed.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3291 de43f980-851b-0410-b2f7-c40aca1f87e0