Add paths to \ior_open:Nn file names
This alters the behaviour of various other parts of the code,
in particular it breaks the \ior_if_eof:N tests as they assumed
we would open a non-existent file. The later is better done
using \file_if_exist:n!

Further changes are required as Bruno has pointed out to allow for
"false" branches and so on.