Overhaul of l3regex to add \regex_show.
Previously, the regex variables consisted of some cryptic code (states 
of the underlying NFA). Now it has more structure, hence is amenable to 
being analysed, letting me write \regex_show:N/n. Besides, the code is 
easier to follow (not yet documented properly). Unfortunately, it also 
means that using a compiled regex is not as quick, since it must still 
be converted to an NFA. I might add some state-related information to 
the compiled regexes to alleviate that.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3381 de43f980-851b-0410-b2f7-c40aca1f87e0