Change "signature" to "argument specification" in xparse messages.
As the word "signature" does not appear in the documentation of xparse, 
it seems best to avoid it in the error messages, even though "argument 
specification" is a bit long.

If I understand correctly, "signature" is used in the code for the code 
containing the argument grabbers, so that's different from the arg spec; 
I didn't change that.