Bug fix: repeated groups could not match an empty tl on repetition.
Naively, when matching "(a??)*a" against "aa", we enter the group, 
match an empty tl (since "a??" prefers matching nothing to matching 
"a"), see the "*" operator, hence try matching the group; again, 
the group matches the empty tl, etc. The loop must be broken. 
Previously, I simply discarded threads when they went back to a 
state that was seen at this position in the string. This is wrong 
if a thread collides with itself: a symptom is that (a??)* should 
always return an empty submatch, but didn't.

Anyways, all is well now, except that the code is not cleaned up.