Also show the value of registers in \tl_show_analysis:n.
Some primitives such as "\boxmaxdepth" etc. have values that could 
in principle also be shown too, but there are too many cases to 
make this very appropriate I think.

The longer-term plan is to have

    %<assert> \assert:n { <boolexpr> }

docstrip guards in higher-level modules, with "\assert:n" defined 
to raise an error if the <boolexpr> is not satisfied: it would then 
show all tokens in the <boolexpr> using "\tl_show_analysis:n". 
Showing the values of registers in this context (and others) would 
help debugging a lot.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3417 de43f980-851b-0410-b2f7-c40aca1f87e0