Tidy \xor_OR_grab_some_cols:nnnnn
As written, this is a pretty hard function to follow. The 'natural'
approach would be to use \prg_case_int:nnn, but I suspect the perfomance
would be an issue. An alternative would be to filter out the 'special
cases' with something like

    \bool_if:nTF
      {
        \int_compare_p:nNn \tex_outputpenalty:D < { -\c_ten_thousand } &&
        \int_compare_p:nNn \tex_outputpenalty:D > 
          { -\c_xor_output_collect_penalty }
      }

and then split for different cases. Yet another approach would be to hard
code the various values so a \if_case:w switch could be used.

There is a note here about dealing correctly with the 'special' penalties,
but I am not clear on what actually needs to be done in that regard. I assume
that simply removing the guard box is not correct, as otherwise Frank would
have done it!

There are a couple of "\ERROR..." functions here that I am unsure about.
I have left both in place, but could do with some further information about
both (as marked in the source, this is probably one for Frank).