Replace some fp by arrays internally to allow multi-argument functions.
The goal was to add min and max(x,y,...).  This is now parsed by making 
the comma into a binary operator which "builds" an array using the 
previous floating point and whatever we read next.  The function (min or 
max) does not influence the array being built: hence, (x,y,...) on its 
own is valid.

Arrays are shallow: (1,(2,(3)),4) is the same as (1,2,3,4).  I had to 
decide how binary relation operate on such objects.  At the moment, 
(1,2)*(3,4) is equal to (1*3,2*4), with no error checking on the lengths 
of the two arrays.  Unary functions also map on each item, so sin(1,2) 
or sin((1,2)) is the same as (sin(1),sin(2)).

Once everything is evaluated, the result should be a single floating 
point number, since functions like \fp_set:Nn and friends really expect 
that.  This restriction could be lifted if we introduce a floating point 
array data type.  In fact, there doesn't even need to be a separate 
data-type for "single fp" and "array of fp", since a single fp is simply 
an array of 1 fp.