Swap in new (BLF) FPU to replace old (JAW) one
This commit replaces the non-expandable FPU with the new expandable one.

The FPU is built as a single .sty file for loading with LaTeX2e, and is part of
the standard monolithic LaTeX3 kernel structure.

At present, none of the 'old' functions \fp_<operation>:Nn are included in docs
(i.e. all are in l3fp-old for deprecation). This is not final: decision to be
made by team before next CTAN update.

Test files m3fp-old001 and m3fp-old003 currently do not pass as the output
format of numbers is not identical. This needs to be discussed: acceptable?

Other tweaks, for example to l3coffins, to be done as separate commits.