Fix xor test file
Two issues here:
 - A 'show' used for some ad hoc tracing was not removed before commit
 - Missing space in trace output

The "unused" guards are used here as the \int_gzero_new:c were added in
in the num => int change as precautions, and are now being removed
where they do not seem to be needed. This happens where the variables
are properly declared elsewhere, but is not so easy to trace and so is
done cautiously.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3613 de43f980-851b-0410-b2f7-c40aca1f87e0