Update \box_show:N(nn) and add \box_log:N(nn)
The old version of \box_show:N did not work in the same way as other
...show:N functions, and also was not actually that informative.
In the new version, \box_show:N(nn) writes to the terminal as well as
the log, and \box_show:N outputs the full detail of the box content.

The \box_log:N(nn) functions write only to the log: boxes can be
very complex, and so writing to the terminal is not always the most
useful outcome.