Further tidying of the driver code
It makes sense to cache sine and cosine values within the box rotation
routine, even if the FPU does not, so we retain \l_box_(sin|cos)_fp here.

An expandable FPU makes it easier to write negative values directly
to the output, which is therefore done.

Values are rounded so there is not much difference in the PDF between
the native drivers and the LaTeX2e ones: the extra digits available to
us don't really add any value.