Document false branch assignments as unspecified.
When a branching get or pop function takes the false branch, it is not 
clear whether we should assign \q_no_value to the tl variable or not.  
The current decision is to leave it unspecified.  I figured that 
changing this aspect of the documentation counts as an update to the 
functions, hence changed the "update" date.