Remove use of color package and move loading graphics to end of preamble
Loading graphics can be moved such that we only load if at the end of
the preamble the user has not, unless we use our own drivers in which case
the issue does not arise

Box clipping does not work without the native drivers as graphics does not
implement ti (only clipping images): an appropriate message is now issued

Loading color is dropped entirely, which highlights an issue with how
vertical coffins were implemented. The \vtop primitive gives the wrong
result if there is a whatsit at the start of the box. So
\color_ensure_current: should not be added here. The test files change as
a result.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3672 de43f980-851b-0410-b2f7-c40aca1f87e0