Enable xparse-defined environments to treat starred version as a parameter
With this change, using

  \DeclareDocumentEnvironment { foo } { s } ...

will mean that 

  \begin{foo*}

is equivalent to 

  \begin{foo}*

without needing two separate definitions for foo.

At present, this fix works only in package mode. In format mode, it will
probably be more sensible to handle this when an environment is used
rather than when it's defined (less wasteful in csnames). However, patching
the LaTeX2e kernel to alter the behaviour is far too risky to be the
best approach for this case.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3684 de43f980-851b-0410-b2f7-c40aca1f87e0