Parse integer registers properly, not through a dimension.
I was using \dim_strip_pt:n { #1 pt } for all registers "#1".  This of 
course will fail for large integers.  The correct way is to apply 
"\the", then grab everything until the first "pt".

Incidentally, this won't fail with a toks register if it contains 
something sensible, but it will fail when \count, \dimen, etc are used 
with an explicit number.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3784 de43f980-851b-0410-b2f7-c40aca1f87e0