Update l3box to use new "@@" mark up for internal material
Each module has different issues to consider when making these changes, for
example there are a number of 'borderline' cases where the interal/documented
status is not clear or where minor errors in the documentation become apparent
when marking material as internal. As such, it seems best to do a small number
of modules per check in.

Notes for l3box:
 - \box_log:Nnn was not properly marked as documented: this is now corrected



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3789 de43f980-851b-0410-b2f7-c40aca1f87e0