Using an fp directly now leads to an error rather than it being typeset.
Since fps are more like registers than like token lists, it is better to 
produce an error when such a variable appears away from an fp function. 
To reflect this, the name of an internal building block of floating 
points (which now triggers the error) has been updated:

  \@@_use:w => \@@_chk:w

The names \@@_error:w or \@@_misused:w would perhaps be better suited, 
but since this will appear in the definition of every fp variable, I'd 
rather not worry people that their fp has an "error" or is "misused" 
when they look at the \meaning.