Rename \iow_wrap:xnnnN to \iow_wrap:nnnN
This clears up the fact that the x-type expansion here takes place
_internally_, not as part of the argument passing. The documentation has
been revised to explain this carefully (I hope).

At the same time, the 'run-on count' argument has been dropped: this can
simply be calculated from the 'run-on text'.

The next step is to sort out the x-type arguments in l3msg: done separately
to keep the commits clear


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3861 de43f980-851b-0410-b2f7-c40aca1f87e0