Get l3doc to infer '@@' replacement from the docstrip lines.
l3doc now does the same replacement as l3docstrip, within
the macrocode environment, in the argument of the macro
environment, and the argument of \cs (am I missing any?).
The module name to use in replacing '@@' is automatically
obtained from searching for '<@@=...>' in the contents of
the macrocode environment.

This means that all modules are now typeset properly.

There are problems with the \label and/or \pageref used to
cross-link from implementation to documentation.  Macros
with an empty signature fail, and macro blocks where the
last macro is not a base function (i.e., most of the cases)
fail too.  I think those problems are unrelated to this
commit.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3906 de43f980-851b-0410-b2f7-c40aca1f87e0