Improve interaction between @filelist and our recording mechanism.
LaTeX2e's \@addtofilelist is redefined to be \use_none:n at the start of 
the document.  Using that exclusively meant that the names of files 
loaded after that were lost (as they are in LaTeX2e).  Also, \file_list: 
must take \@filelist into account if it is not yet stored in our 
recording sequence.  There is an additional subtlety about when file 
names are strings or not.

Minor doc tweaks.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3910 de43f980-851b-0410-b2f7-c40aca1f87e0