Expandable warnings.  Warning about deprecated functions.
Abusing some eTeX warnings (when tracingnesting > 1), we can obtain 
expandable warnings.  This means that both expandable and non-expandable 
deprecated functions can produce a warning at point of use.

At the moment I just modified a bunch of l3kernel deprecated functions 
by adding \__cs_deprecated:N <function> to their definition.  If we 
decide to go with that idea, we should put that in the definitions in 
the first place.

This commmit probably also contains a bunch of additions to the 
'miscellaneous' section, which come from various unresolved email 
threads, and don't need to be resolved any time soon.