Correct precedences in boolean expressions (Fixes #58)
I marked as updated all functions which take a boolean
expression argument.  Not sure that's the right thing
to do.

The fix is ugly and slow, but at the moment I can't think
of something better.  Replace all '(' by '((', all ')' by
'))', all '||' by ')||('.  This gives the expression the
form

   ( (a && b && ...) || (c && d && ...) || ...)

which means that we can use the already existing parser
without any change, since a given group will either
contain only && or only ||.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3927 de43f980-851b-0410-b2f7-c40aca1f87e0