Remove \__prop_split:Nnn.  Alter prop_put (order of keyvals)
The first change avoids a relatively slow \exp_args:Nno which appeared 
in \__prop_split:Nnn (only used by the user function \prop_put:Nnn).

The second change is as follows.  Previously, if the key was already 
present in the prop, it was removed, and the new value put at the end of 
the list.  Now the new value is inserted at the place where the key was 
originally.

The order of keys is thus more stable (if I am not mistaken, keys appear 
in the property list in the order in which they are added to the prop), 
which should allow us to provide an 'array' data type, based on props 
with integer keys, and with fast looping through items _in order_ 1, 2, 
3, ... With props as implemented earlier, the order would not be 
preserved.