Revert 'auto-magic' "\begin{foo}*" <=> "\begin{foo*}" mapping
This change was added in SVN3684, but the team as a whole are unsure that
it is the correct approach to deal with this issue. As such, further
discussion is needed before any public release of such a change.

Note that SVN3684 also included a change in the end-of-group code
for environments in IniTeX mode. This is left in place here, as it is
required irrespective of the handling of *-environments. 


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@3956 de43f980-851b-0410-b2f7-c40aca1f87e0