Revert 'auto-magic' "\begin{foo}*" <=> "\begin{foo*}" mapping
This change was added in SVN3684, but the team as a whole are unsure that
it is the correct approach to deal with this issue. As such, further
discussion is needed before any public release of such a change.

Note that SVN3684 also included a change in the end-of-group code
for environments in IniTeX mode. This is left in place here, as it is
required irrespective of the handling of *-environments.