New 'xbox' trial module
The aim here is primarily to explore how well the 'coffin' concept works for
producing document-level boxes in the way LaTeX2e does. At the same time, there is
the opportunity to also to see how well different 'layers' work out with xparse
when implementing more complex functionality.

The general impression is good: even lrbox was not too hard in the end. Whether
coffins-as-boxes at the document level makes sense I am still no sure, but
if we for example extended \usebox then this might make sense.

At this stage, most of the hbox-based material matching LaTeX2e's ltboxes.dtx is
implemented. The missing hbox functions are those involving frames: I am
still working out how to deal with the fact that \frame adds the frame 'outside'
of the material but \framebox adds it 'inside'. (The problem is coming up with
a code-level interface to support the document-level one: the actual
implementation is fine.)

The interfaces implemented at present are exactly those in LaTeX2e. You can
imagine adding a keyval interface to several functions using the now-standard
idea of a first optional argument. This might make for example \makebox a bit
easier to follow. 

Not considered at all at this stage are the vbox-based structures (\parbox,
minipage), as these will involve some galley questions.

There are some 'odds and ends' in ltxboxes that I am not sure where to place.
For example, \llap is not documented in Leslie's book, but is probably wanted
at the document level for LaTeX3 and seems sort-of box-related. On the other
hand, something like \rule feels like it should be in a separate module, as
does \underline, while \centerline, etc., perhaps should simply not exist at all.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4024 de43f980-851b-0410-b2f7-c40aca1f87e0