Add a code-level version of \unitlength
Thsi fixes the code which converts from (x,y) to dimensions when used in
iniTeX mode: the old version was broken. Of course, this is not a particularly
good way to set this up (it's not part of the xbox module really), but for
testing it should at least work.

Open question: what do we want to do about picture mode? I can see an argument
for going for a pgf/TikZ like approach to this entire area, in which case
the syntax is not that helpful. On the other hand, if we want to retain
everything in Leslie's book (do we?), then we need it even if it's marked
as deprecated. Worth noting that TikZ also has the same unit length concept,
so that part will stay somewhere.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4038 de43f980-851b-0410-b2f7-c40aca1f87e0