Try to clarify what happens to spaces, tabs and trailing newlines.
Unfortunately, there is no way to avoid losing spaces and tabs at the 
end of each line, and a trailing newline at the end of each file, since 
TeX does that in its internal file-reading mechanisms.

The file contents can be read exactly using pdfTeX's \pdffiledump 
primitive, or a Lua equivalent in LuaTeX; XeTeX is unable to do that.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4053 de43f980-851b-0410-b2f7-c40aca1f87e0