Add l3trial/l3check, with only l3expan checks so far.
The goal is to go through the kernel modules one at a time,
trying to redefine functions to do as many checks as possible.
The current code only redefines the l3expan internals, which
then make sure that, e.g., \use:c raises an error when it
builds \relax; or \::o and others raise warnings when their
argument does not change under that kind of expansion.

Loading l3check after l3candidates in expl3 leads to a couple
of failing tests, which most often point to not-so-good code
or weird tests (see the past few commits).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4075 de43f980-851b-0410-b2f7-c40aca1f87e0