Some initial thoughts on re-implementing \framebox
This is all something of a pain, as LaTeX2e has two ways of framing material.
\framebox adds a frame such that the resulting box expands, whereas \frame
and the picture-mode version of \framebox add the frame 'around' the content.
I am not convinced I've got the interfaces right yet.

The LaTeX2e implementation for framing draws the frame in a somewhat haphazard
way, as one rule is behind the content and the others are in front. I have
altered things such that the frame is 'on top' on the content.