Add first testfiles to xbox
These cover the various 'save box' and 'make box' functions at the document
level, but not 'frame box', anything TeX-derived (e.g. \llap) or anything
code level.

To get closer to what LaTeX2e gives, I have altered the interfaces of
\savebox and \makebox such that with no optional arguments they do the same
as \sbox and \mbox, respectively. This avoids adding additional boxes or
glue to the output. Quite what is 'expected' here I am not sure: perhaps it
should be the other way around and \sbox/\mbox should be altered to always
do the full internals. (Issue: \mbox is used a lot, and there would be a
performance hit for no actual gain.)