Add l3check code for l3expan, l3tl, l3seq
I've been oscillating quite a bit between different interfaces, so the 
l3check code is a mess.  It can be loaded with \RequirePackage{l3check}, 
and loading it with the kernel testfiles lead to a bunch of commits 
earlier today, and finding a couple of bugs in l3kernel.  This still 
needs lots of work, but it has low priority for now.