Store number of keys in datatable header
To improve dt performance, there will be a series of changes. The first is to
store the number of keys (columns) in a table. This will be used later to
allow storing columns by number rather than by name. The idea here is that
this cuts down on the total number of tokens in the table, which has a
performance impact. Datatool uses a similar approach, and there is a clear
benefit in speed.