Trim overcomplete variants (fixes #100), detect o => x, and allow o => o
With a base form "nnTF", requesting the "xx" or the "xxTF" variant is 
equivalent (more generally, any trailing common string is ignored).  

With a base form "xn", requesting the "nn" variant is an error, because 
there is no way to go from x-expansion to n-expansion (more generally, 
only n and N arguments are allowed to change when creating variants). 

With a base form "on", requesting "oo" or "ox" (etc) is allowed, and 
leads to using "\exp_args:Nno", correctly avoiding any expansion on the 
first argument, since the base function already uses the letter "o".

Hopefully, this does not come with too much of a performance penalty: 
loading expl3 seems to take just as much time.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4147 de43f980-851b-0410-b2f7-c40aca1f87e0