Change mechanism used to dispatch when finding \relax in fp expressions.
When the fp parser sees a cs equal to \relax, it can mean three things: 
(1) the cs is \s__fp, and is followed by an internal floating point 
number; (2) the cs is \s__fp_mark marking the end of the expression, 
this triggers a "premature-end" error; (3) anything else is interpreted 
as a badly initialized variable.  Dispatching with 
\__fp_type_from_scan:N is a bit slower than the custom auxiliary used 
before, but on the longer term it will allow to add datatypes.