A first go at support for symbolic expressions.
There, I'm not aiming for speed, but aiming to make the code as short as 
possible.  Expressions are evaluated as much as possible, stopping when 
an unknown is encountered.  The current implementation has the same 
infinite-recursion problem as conventional macros: the simplest example is

    \fp_var_set:nn { A } { A }
    \fp_eval:n { A }

This can be resolved by letting previously undefined variables to NaN 
(or similar) before evaluating the expression, to avoid such 
self-references.