Add f-type expansions tests for \tl_head:n/\tl_tail:n
These tests fail with l3tl.dtx 'as is', as the current defintion is not
f-type expandable. They all pass with a version of \tl_head:n proposed by
Bruno:

    \cs_new:Npn \tl_head:n #1
      {
        \etex_unexpanded:D
          \if_false: { \fi: \__tl_head_aux:nw #1 {} \q_stop }
      }
    \cs_new:Npn \__tl_head_aux:nw #1#2 \q_stop
      {
        \exp_after:wN \__tl_head_auxii:n \exp_after:wN { \if_false: } \fi: {#1}
      }
    \cs_new:Npn \__tl_head_auxii:n #1
      {
        \tl_if_empty:oTF { \use_none:n #1 }
          { #1 }
          { \if_false: { \fi: \__tl_head_aux:nw #1 } }
      }

and all of the f-type tests (but not the \q_stop tests) pass with 

    \cs_new:Npn \tl_head:n #1
      { \etex_unexpanded:D \exp_after:wN { \tl_head:w #1 { } \q_stop } }

(the 'original' version).

(Untested, but they should also pass with David's idea 
http://tex.stackexchange.com/a/69877, which is a loop-based approach and so
rather slow.)


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4198 de43f980-851b-0410-b2f7-c40aca1f87e0