Add an initial 'l3prefixes' tracking list
As suggested on LaTeX-L, the idea here is to have a place to track prefixes
used in LaTeX3 code so that the namespace is at least in some way 'mapped'.
At this stage, this file is experimental, so it's not added to those which
go to CTAN, but presumably will if the idea takes off.

Initial content covers only team code, and only that which goes to CTAN.
The coverage of l3kernel/l3packages/l3experimental should be more or less
complete. There are a few awkward 'primitives renamed' (e.g. \or:,
\reverse_if:N) which are not included. In terms of 'module', the scheme here
is to recongnise that some modules fall into larger bundles, particularly
l3kernel. At the same time, we use module prefixes in some 'cross module'
places, for example in the coffin code (split between l3kernel and xcoffins).

I've used 'homepage' and 'issues' for web pages, but 'repo' for something
that can actually be checked out, rather than a page pointing to a repo. Does
that make sense?

Next additions will probably be 'reserved' prefixes for scratch use: for
example, 'tmp' and any derivatives (tmpa, tmpb, ...) should be marked as
available for use in local code only.

For third-part code (including from team members) I suggest we sort out the
submission system as mentioned on the list, then add from that. I'd expect to
start that ball rolling with siunitx/notes2bib, to give everyone a model to
follow. 

Feedback on all aspects welcome!


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4238 de43f980-851b-0410-b2f7-c40aca1f87e0