Use a single 'block' for reserved registers
As mentioned in issue #221, we can avoid a loop here by using a single
block. (As a side effect, l3alloc can then be loaded a little later if
desired.) The allocation might still be a little faster if we use offsets
rather than absolute register numbers, but this approach is probably
clearer.

It would be nice to have a single set up for the reserved values, but
\tl_map_inline:nn is not available as it requires an int register. So
this is coded by hand: hopefully acceptable here!