Alter l3drivers DocStrip guards to generate a dvipdfmx driver (see issue #119)
This check-in generates two essentially identical drivers: the code for
dvipdfmx is the same as for xdvipdfmx (XeTeX). This is a first step:
testing is required to see how well this works! The two drivers are kept
separate as at some future stage we may need other facilities which will
depend on the engine (the dvipdfmx can use some pdfTeX primitives that
the xdvipdfmx one does not have access to).