Use l3keys to process expl3 options as far as possible (see issue #124)
This check-in uses l3keys to process the options which can be set after
the kernel code itself is loaded. Options "check-declarations" and
"log-functions" are still handled using the 'fake keyval' approach as this
allows them to be used as part of the load process. As Bruno suggested, these
options are also present in the keyval setup so that any typos in the value
parts are trapped by l3keys in the correct way.