Use an internal box for coffin rotations
Issue #126 prompts me to look at this (it still needs fixing). Broadly, the
behaviour of \bo_set_(ht|dp|wd):N is non-standard, so it is safest not to use
them directly for any stored box.

At the same time, there were two kerns in a row which can (hopefully) be replaced
by a single one with no loss of functionality.