Ensure boxes always have the same form after coffin scaling
In general, the box-related code does not take 'short cuts', so any box operation
always produces the same number of box levels, kerns, etc. This was not the case
for coffin scalings: now sorted.

At the same time, the test file is updated for the combination of two kerns to
a single kern from an earlier check-in.