Allow "driver" option to deal with all cases properly
This removes the need for a separate "native-drivers" option (which is
retained for compatibility). It also adds sanity checks to the driver
choices, and allows us to extend the mechanism as needed (only possible
in DVI mode). As such, it closes issue #119 and issue #124: options are
now proper keyval ones, and dvipdfmx is covered.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4297 de43f980-851b-0410-b2f7-c40aca1f87e0