Add driver-dependent test files (fixes #127)
This is something of a brute-force approach, as the test file names are
hard-coded into our scripts to allow the correct engine selection. Ideally,
this information would be in the files themselves (arara-like).

Note that the driver `.tlg` files currently have to be produced by hand.
If they change a lot, this will presumably need to be addressed!


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4301 de43f980-851b-0410-b2f7-c40aca1f87e0