Add driver-dependent test files (fixes #127)
This is something of a brute-force approach, as the test file names are
hard-coded into our scripts to allow the correct engine selection. Ideally,
this information would be in the files themselves (arara-like).

Note that the driver `.tlg` files currently have to be produced by hand.
If they change a lot, this will presumably need to be addressed!