Revise handling of frame width/border
It's not clear to me how things like \fboxrule should be handled at all. So
for the present stick to the least complicated approach: just a straight
dimen. Potentially this should be governed by some key-value interface,
perhaps as a template.

At the same time, \fbox and \framebox where missing from the documentation:
this is now corrected.