Revise handling of frame width/border
It's not clear to me how things like \fboxrule should be handled at all. So
for the present stick to the least complicated approach: just a straight
dimen. Potentially this should be governed by some key-value interface,
perhaps as a template.

At the same time, \fbox and \framebox where missing from the documentation:
this is now corrected.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4315 de43f980-851b-0410-b2f7-c40aca1f87e0