Remove \kern:n
As discussed on the ream list some time ago, to fit with the rest of expl3
there should be separate kern functions for horizontal and vertical mode. On
the other hand, that's not how kerns work so it seems best to simply omit this
here (xbox is mainly about other things, so the discussion of kerns might get
lost).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4316 de43f980-851b-0410-b2f7-c40aca1f87e0