Extend \xbox_frame_to_wd:
The previous definition simply assumed that the border and frame widths
were available. That was not consistent with \hbox_frame:nnn, and also was
broken as an earlier check-in removed the code-level names for \fboxrule
and \fboxsep!

Quite how to handle the border/rule widths is not yet clear, but presumably
the xbox function and the hbox function should be consistent.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4320 de43f980-851b-0410-b2f7-c40aca1f87e0