Add ifpdf to expl3 tests
The option changes earlier mean that at the moment the ifpdf package is always
loaded. This shows up in the log files in a system-dependent way, so load it
before the checked part of the log. Longer-term, we want to have our own test
here and this will be removed, but at the moment it needs to be right.