Update \xbox_leave_vmode: again
Using the primitive conditional improves efficieny here.

The argument for this implementation rather than

    \box_use:N \c_empty_box

(the plain/LaTeX2e approach) is that using the \indent primitive does not
leave any additional tokens in the wrong place relative to those from the
\everypar primitive. This only works with e-TeX available, as it requires
the \leavevmode-like macro not to expand when use for example as the first
token in an \halign cell.