Make setting flag for parshape local
The approach taken in l3galley is that settings which should respect groupings
should be local, meaning that it's mainly galley internals which are global.
The parshape flags were a mixture, but logically all should either be local
or global. The latter would break grouping, so go with all-local.

There are various other parts of the current l3galley which need checking in
this regard: expect more check-ins over the coming days.