Terminate parshape assignments cleanly.
There were some (harmless, it turns out) "loose ends" where
the last dimension in a parshape assignment was explicit and
not terminated with a space.  I decided to use \exp_stop_f:
as a space to end such dimensions rather than \c_space_tl
because the expansion that TeX performs there is precisely
f-expansion.  Also, that's a bit more efficient.