Improve management of \prevgraf info, and fix 'multi-par' shapes
The current conceptual paragraph may contain one or more \par primitives
(for example, to create a display area). As such, the galley data on the
number of lines in the previous paragraph cannot just be \prevgraf. At the
same time, this also means the current partial paragraph has to be tracked
separately from the previous one. This check-in fixes the order of resetting
of this information so it actually works.

At the same time, the code for setting a shape which applies to a fixed number
of lines and can extend past a paragraph boundary was wrong. The appropriate
code is now used whenever a shape has to be set.