Probe the behavior of peek functions by adding tests
Particularly interesting is the fact that the current peek_charcode 
treats expandable active characters as having their catcode but 
non-expandable ones as being control sequences.  On the other hand, 
peek_catcode treats any active character (not let to a character token, 
because that's a whole new can of worms) like a control sequence in all 
cases, which seems easier to document.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4379 de43f980-851b-0410-b2f7-c40aca1f87e0