Add a cutout system to the galley
This set up separates out cutouts (apply to a fixed number of lines) and
paragraph shapes (apply to one or more paragraphs but do not depend on
number of lines). Cutouts apply globally within a galley and are additive,
with the 'used' lines removed at the end of a paragraph.