Change internal structure of props, removing a trailing q__prop
I had mistakenly thought earlier that property lists had to
take the form
    \q <key_1> \q {<value_1>} ...
    \q <key_N> \q {<value_N>} \q
with a trailing \q__prop, and in particular that empty props
had to be non-empty for performance reasons.  This turns out
to be incorrect, and the new code has similar performance,
with the added benefit that more macros are identical to the
token list ones.

All functions that are directly or indirectly affected by this
changed have been affected a new "updated" date.  Perhaps that
choice is not good, I can revert it if needed.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4401 de43f980-851b-0410-b2f7-c40aca1f87e0