Change \__prop_split:NnTF, less prop auxiliaries.
The T and F branches of \__prop_split:NnTF now receive
the various pieces of the prop as #1, #2, #3, rather than
as three brace groups.  This turns out to require less
auxiliaries, but more importantly, arguments are read one
time less, which leads to speed ups for large props (but
a slight slow down for small props).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4423 de43f980-851b-0410-b2f7-c40aca1f87e0