Fix bug in \str_substr:nnn, fix documentation.
The start and end index of the substring must be clipped to the size of 
the string.  This was done incorrectly, and meant that even for a very
large start index, the result was not empty.

The behaviour when the end index is 0 has changed to match the case of 
an empty end index (both because it seems more natural and because it is 
easier to code).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4440 de43f980-851b-0410-b2f7-c40aca1f87e0