Strip exactly one set of braces if present inside optional arguments
This is what LaTeX2e does, but is done here in a very carefully controlled
way (so that the other ideas we have in xparse work).

Note that this may well be a rather brute-force approach to solving the issue:
there is probably a better way to implement all of this!