First version of suggested \tl_(lower|upper)case:nn (see #141)
I wonder if we might actually be better if the first argument
is a mapping

    \tl_lowercase:nn
      {
         { `\A } { `\A }
         { `\B } { `\c }
      }
      { <stuff> }

but then this has the issue that there are different ways of 
giving a charcode and if you go with the above you can only accept
one (presumably as I've done above by number).

Longer-term, the above might suggest we don't need `\char_set_lccode:n`,
etc., at all, but I'm wary of that as there is an interaction between
lccode/uccode and for example end-of-sentence spacing.

Thoughts on all of this most welcome!