Update clipping in (x)dvipdfmx driver
When PostScript specials are mixed in with PDF ones, for example if using
the LaTeX2e `graphics` package to scale content, the previous version of the
clipping code gave incorrect results. This seems to be because `dvipdfmx'`s
`pdf:bcontent`/`pdf:econtent` pairing only places material correctly if all
specials involved are `pdf:`-based.

The updated approach does the correction of position manually, such that
the clipping path and box insertion will always be in the same place
irrespective of the content of the box. The trick is taken from the `pgf`
`xdvipdfmx` driver. 


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4505 de43f980-851b-0410-b2f7-c40aca1f87e0