Extend key property variants
As these are not really easy to do using standard tools, we need a more-or-less
complete set. I've left `\generate_choices:n` as this is likely to be
deprecated. Note that for code arguments we have n/x versions only, while
for 'general stuff' I've gone for n/V/o/x, as f-type seems relatively
unlikely.