Add tests for .meta:n and .meta:nn, and remove variants of these
Adding the tests, I found that a 'proper' .meta:x is actually not what
the earlier .meta:x was doing. Instead, it was essentially .meta_x:n. That
may be useful, but I will add only if required. Currently, the only place
in TeX Live where .meta:x is used is in my own code (notes2bib/siunitx). I
will therefore look to fix them and see if this requires a .meta_x:n
property.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4538 de43f980-851b-0410-b2f7-c40aca1f87e0