Remove .code:x and related properties
After a suggestion from Bruno, I checked on use of these: not in use by
anyone. Moreover, looking at the inspiration in pgfkeys they are used there
only as it uses keys to define keys (OO), which we do not. Finally, and most
importantly, the x-type expansion here was a bit odd as it was not the usual
\cs_generate_variant:Nn code. 

Since no-one is using these, I think its safest to simply drop (as they are
actually 'not quite right' anyway).