Use a single property list for key information
The upcoming idea of 'key groups' means we have at least three things to
track (default, required/forbidden and group). A property list makes a lot more
sense than lots of tls. A few internal changes are needed to cover the fact that
using a prop is a slightly different thing as we need to watch whether it 
actually exists.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4546 de43f980-851b-0410-b2f7-c40aca1f87e0