Accelerate prop_map_inline.  Define s_obj_end.
Since property lists are now, up to two tokens which are
equal to scan_stop:, a list of pairs of the form

    \__prop_pair:wn <key> \s__prop { <value> }

we can map a function through such a list quite quickly
(but not expandably), by defining \__prop_pair:wn.  This
is exactly the same idea as \__seq_item:n for sequences.
This commit requires giving a definition to s_obj_end,
which I put in l3prop.dtx, for no good reason.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4563 de43f980-851b-0410-b2f7-c40aca1f87e0