Change internal structure of sequences.
In order to experiment with some l3obj ideas, I need the already
existing complex types (props, seqs, at least) to have a very
particular form, hence the recent changes.  Contrarily to the
prop changes which have no performance impact whatsoever, I am
not quite sure about this set of changes (no huge performance
hit, though, of course).  The main culprit is the trailing token
which I insist on having on all sequences.  It might turn out to
be useless, in which case I will delete it and simplify the code.