Remove deprecated functions which are not DocStripped into .sty files
We are approaching the point where no deprecated functions are removed from
the code, but are simply not used or documented. As such, it seems sensible
to remove those older deprecated functions which are not available in the
.sty files. All of these are documented in l3obsolete, with removal dates and
check-in numbers, so can be recovered from the repo if required.

Note that three functions which were meant to have been removed earlier
had 'escaped', and are removed here.