Change pi, argument reduction, rules for juxtaposition and for dim registers
Argument reduction for trigonometric functions in radians (sin, cos, tan, cot,
sec, csc) now uses a very precise approximation of 2pi, so that the result is
correctly rounded even for very large arguments (the current maximum exponent
is 10000, so we need a bit more decimals of 1/(2pi)).

The weird rounding previously used for pi is abolished: pi is now
3.141592653589793 (was ...4).

Previously, (3 + 4)pi was correct but 3(4 + pi) led to a 'Missing * inserted'
error.  Now both are allowed, as is (3 + 4)(5 + 6).

Dim registers were converted to floating points through TeX's primitive
\tex_the:D, and we now use \dim_to_fp:n, which rounds the exact number
(number of sp)/2^{16} to 16 significant digits.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4601 de43f980-851b-0410-b2f7-c40aca1f87e0