Use sind and cosd.  Round angle/sin/cos to 5 digits, then compare to 0
The angle for box rotations is given in degrees: the new sind and cosd
are faster and more accurate than multiplying by deg and calling sin
and cos.  The driver code checks for zero to avoid writing '-0' in
specials ('-0' confuses some viewers), but the comparison should be
done after rounding to 5 places since rounding can result in a zero.
This fixes the failing test m3box004.