Support user-defined fp functions.
This commit does not provide ways to define new 'words'
such as 'sin' or 'exp', but to define control sequences
which behave the same as those words.  See test file
l3kernel/testfiles/m3fp-func001.lvt for some examples.
The feature may be removed if it turns out that it does
not fit.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4618 de43f980-851b-0410-b2f7-c40aca1f87e0