Implement square root with correct rounding as sqrt.
We cannot always guarantee correct rounding (i.e., what would
be obtained by rounding the exact result to a representable
number) when raising a floating point number to the power of
another such as x^0.5.  For the specific case of square roots,
this can be done, and is required by IEEE-754.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4619 de43f980-851b-0410-b2f7-c40aca1f87e0