Implement square root with correct rounding as sqrt.
We cannot always guarantee correct rounding (i.e., what would
be obtained by rounding the exact result to a representable
number) when raising a floating point number to the power of
another such as x^0.5.  For the specific case of square roots,
this can be done, and is required by IEEE-754.