Correct parsing of multiple comparisons in a row
Previously, (3 > 2 > 1) was parsed as ((3 > 2) > 1), which
is 'false'.  Now it is parsed as (3 > 2) && (2 > 1), which
is 'true'.  The new behaviour is consistent with the
integer function int_compare:nTF and others.  However, lazy
evaluation is not supported, i.e., even if an inequality
evaluates to 'false', all further calculations will be
performed before returning 'false'.