Remove trailing markers from seq and prop structures
When adding trailing markers earlier on, my thought was
that they would be needed if objects were to be used, for
instance, within floating point expressions.  The trailing
marker come with a performance hit (10%? 20%?) which is
noticeable when compiling source3.

A better approach will be that any object which may appear
within floating point expressions should be marked as such
and thus be packed into something like

   \s__fp_obj { the actual object }

thus allowing to grab the object all at once rather than
having to guess where the object end (remember, the floating
point expression parser expands everything, so being stored
into a macro does not help).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4623 de43f980-851b-0410-b2f7-c40aca1f87e0