Update driver code to correct issues with (x)dvipdfmx output (fixes #156)
This update reworks quite a bit of the driver file. This is partly to sort
out the problem linked and partly to address other issues. The specific
problem was due again to the need to allow (x)dvipdfmx to deal with both
PostScript and PDF material. The approach taken now is the same as in the
graphics driver for XeTeX: use raw q/Q save/restore for the clipping code.

The wider changes tidy up various bits of the drivers, and also reflect the
fact that the amount of common code realistically managable is small.