Tighten up \pdfstrcmp emulation/testing
When LuaTeX is in use in package mode we should not simply copy \pdf@strcmp
as \pdfstrcmp: there is a reason HO doesn't do that in the pdfstrcmp package!
Instead, test properly for both \pdfstrcmp and \pdf@strcmp with the appropriate
adjustment in l3names to pick this up. In format mode the problem is not there:
as all of the primitive names are removed anyway and as we already take care
to allow for \pdftex_strcmp:D being 'emulation', there is no need to worry.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4640 de43f980-851b-0410-b2f7-c40aca1f87e0