Fix \seq_mapthread_function:NNN
Bruno's ealier checkin (4623) removing the 'end of object' markers was
a bit over-zealos, and here some form of 'end of sequence' marker is 
required to do the expansion properly and allow trimming off of the first
token in each case (\s__seq). As a fix at present, basically go back to the
older method using \q_stop to do the job: this may need to be improved later.

Note: A test file is needed to cover this. I will look at this a little later
on.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4643 de43f980-851b-0410-b2f7-c40aca1f87e0