Document treatment of trailing spaces by \tl_rescan:nn, etc. (closes #154)
The e-TeX \scantokens primitive treats its argument as a file, and so spaces
at the end are dropped. Adding a marker token at the end of the input can be
used to avoid this but at the cost that this token cannot appear in the real
input. As such, that is not a general fix and instead it's best to document
what happens here.