Set options one-by-one in l3keys2e (fixes #161)
To allow \CurrentOption to be defined for use with LaTeX2e, the options
need to be available 'before' key processing. Without hacking into l3keys
internals, the only way to do that is to do a clist mapping and only set the
keys one-by-one. Performance hit here should not be major in the grand
scheme of things.