Allow for nesting in \keys_set_known:nnN
The approach taken here is much the same as is used for saving module
path in \keys_set:nn (which itself is from pgfkeys). The current content
of the 'saved' clist is stored in the input stream and 'put back' after the
selectie setting. The result is that the 'internals' of the setting system
do not need to know anything about the variable used: they can all stick
with a single scratch one.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4705 de43f980-851b-0410-b2f7-c40aca1f87e0