Enable 'generic' loading of expl3/DocStrip most of expl3 into one file
This is a 'first pass' in getting everything working: at present it should
function correctly with LaTeX2e (of course), plain TeX and ConTeXt MkII. There
are a few issues outstanding for ConTeXt MkIV: I expect to address those later
today in a series of smaller commits.

The overall plan here is to DocStrip most of the code layer into one big
'generic' file and to provide two small loaders, expl3.sty (LaTeX2e) and
exp3.tex (all formats). The latter loader is currently designed to detect
LaTeX2e and to 'switch' to the .sty route. Feedback on file names, etc.,
welcome.

Rearranging the DocStrip guards has meant moving some code from expl3.dtx to
other places. There are a few parts of this that need tidying up: to do
over the coming day or so.

The split between expl3.dtx and l3bootstrap.dtx is now a bit clearer: expl3
contains the 'loader' code (different between expl3.sty and expl3.tex),
while l3bootstrap contains stuff that is common to both (and also initex mode).

Removing l3bootstrap has a knock-on across our code but should not affect
anyone else: in TL2014 l3bootstrap is only loaded by the expl3 modules.

As well these changes, there are a few fixes to the documentation so that
source3 builds (needed to test it works).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4712 de43f980-851b-0410-b2f7-c40aca1f87e0