Use private version of \pdfstrcmp with LuaTeX andrevise loading of etex/luatex
Using the pdftexcmds package for \pdfstrcmp emualtion with LuaTeX causes issues
with ConTeXt. It also is an avoidable dependency, so is dropped here. The
code for emulating the primitive is tightened up: open question as to whcih Lua namespace we shoud use for this!

Other support package loading is rearranged to be loaded only once the test
for the required primitives is done and only for formats where this is
necessary.