Update make.bat to reflect 'generic' mode changes
As there is no longer a separate l3names.sty, the test file m3tokens004
now needs some other way of getting a list of primitives. This information
in more-or-less the same format as was in l3names.sty is saved in names.tex:
this needs to be appropriately copied by make.bat.

Also there was a typo in make.bat concerning "make clean": was removing
some files it should not! 


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4720 de43f980-851b-0410-b2f7-c40aca1f87e0