Improve internal logic in low-level key parser
As a result of the issue with keyval and brace stripping, I've taken a quick
look over the implementation in l3keys for the same low-level task. No issues
with the defined behaviour in l3keys but the logic was not so clear. By adding
in one extra quark we can use much faster tests for the branching needed, which
makes the code more efficient and also clearer.



git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4752 de43f980-851b-0410-b2f7-c40aca1f87e0