Set up to run engine-dependent tests
The logic here is as follows. Each test requires one .lvt file and at least
one .tlg file. For a <test>, the file <test>.tlg is used as the comparison
unless on called <test>.<engine>.tlg is available. The base <test>.tlg
files are created by pdfTeX, and so the engine-dependent ones will be
required for XeTeX/LuaTeX at present.

This checkin does not include any new test files: they will be supplied
separately.

Note that to test the drivers a new approach will be used once l3kernel is
set up. The driver tests can be altered to use

  \ifdefined\pdfoutput
    \pdfoutput=0
  \fi

at the very beginning to run the dvips/dvipdfmx tests, and this will allow
them to run with pdfTeX/LuaTeX.