Add a copy of the file input test file nested 'one below' the current location
The Lua test scripts run 'in place', so this will be the correct place for the
file.

Note: It occurs to me that the renames are probably not needed now that the Lua
scripts run tests 'in place'. However, I think they are still reasonably 
userful: make it clear which files go with what.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4817 de43f980-851b-0410-b2f7-c40aca1f87e0