Improve m3box003 test
The old version didn't actually \box_show:N some boxes, meaning that
what was in the log depended on the setting applicable in regression-test.tex.
That shows up with LuaTeX, as since it puts more stuff in each box there
is an "etc." replacing some content. Moving to 'full' boxes and deliberately
showing means that we are in control with \box_show:N.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@4869 de43f980-851b-0410-b2f7-c40aca1f87e0