Generalise "make cmdcheck"
This can now be run on all modules and with selection of input files:
engine used is "stdengine" (doesn't relly need another var).

As l3doc fails, l3fp fails and the l3fp-... files are a waste of time
I've given a list of .dtx files for l3kernel which skips them.