Set escapechar very locally to avoid LuaTeX bug
LuaTeX's luaescapestring fails when the escape char is negative.
This means that our emulation of pdfstrcmp fails.  Since that is
used to check for the end of mappings, and in particular used in
the implementation of file_input:n, inputting the list of primitives
fails if we have already set escapechar=-1.  Instead, set that in
a group just before doing each test.