Make the documentation of int_step consistent with the code
This change in documentation is independent of the minor
change in code in the last two commits.  The issue here is to
try and clarify when the step and final value are evaluated.
The earlier documentation was wrongly claiming that they are
evaluated up front.  Instead, the code evaluates them at every
step in the loop.