Generalise support files concept
Split into 'unpack' and 'check' versions: this means docstrip.tex is not
hard-coded. As checking is the step most likely to be run 'isolated', the
default for that is 'everything sensible in support'.

I've not yet added a 'typeset' possibility: I have a feeling that the name
of the vars should match the target, so either the target should be "typeset"
or the vars should all be "doc...": thoughts?

I think this allows the cp function to be simplified again: will check and
then commit.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5035 de43f980-851b-0410-b2f7-c40aca1f87e0