l3build: add "typesetsuppfiles" variable
it seems useful to be able to add files to "support/" that can be used in the compilation of doc.
otherwise you'd need to litter the main folder with files for typesetting.

longer term, we might want to consider having a separate build directory for typesetting in which the finished product is moved back into the main dir if desired.

other considerations: typesetruns counter, inclusion of bibtex/biber, glossaries, ... perhaps totally separate build scripts that we can just call externally? Don't want to re-invent the wheel too much in here.