l3build: Typeset docs in a dedicated location
I ran into an issue with the latest CTAN release where a file 'here' was
causing issues. By using a dedicated directory, we control what files are
available to the process: in this case, the local files, source file itself
and then system tree if allowed.

At present, the end point here is still to copy PDF files back to the original
location: seems most useful approach but can be altered if required.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5107 de43f980-851b-0410-b2f7-c40aca1f87e0