Revert r5107 and r5108
Life is actually more complicated than the previous commits suggest. There
is a need to worry about multiple source files, and copying them is required
for in e.g. l3kernel to do the job 'properly'. That then makes life more
complex with paths. For the moment, stick with typesetting in place.