l3build: Add demo Lua scripts directly to .dtx
A few reasons for this. Requiring other files makes life tricky for moving
to an 'isolated' build approach, and ties the file to how our repo is set up.
It also means that to allow inclusion in TeX Live we'd really need to ship the
.lua files to CTAN, which is not really the idea.

Upcoming checkins will reinstate the idea of building docs in a separate
dir: I now have this sorted in my head.