Create new l3str 'kernel level' module
This moves all of the \str_... code into a single module separate from l3basics/
l3tl. No new functionality is added here: purely a rearrangement. At the same
time, so that there is only one l3str.dtx the experimental file of the same
name is renamed to l3str-expl.dtx. As we now DocStrip the kernel into a single
file, there is no issue with name clashes for the .sty side of things.

This check-in is the first stage of adding more str functions, in particular
case-folding.