Add script and extracted data for Unicode case changes
The script here extracts case change data from the various Unicode files. As
this process can't really be done as part of a general build process, the
extracted data itself is also checked in. The script runs with LuaTeX,
tyepsetting the documenation will work with pdfLaTeX/XeLaTeX/LuaLaTeX.

Adding the case change functions themsevles is yet to be done: at present only
string 'case folding' is ready (expect this to appear later today).