Only save case-folding data to .def file at present
As the case-changing stuff is not yet implemented at all it's not really
necessary/sensible to include the data for that part in the .def. (A simple
version can of course be created using the folding code as a basis and omitting
the context-dependent stuff Unicode define.) Namespace for case-changing is
also not decided yet, so avoid making a bad decision at this stage.




git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5150 de43f980-851b-0410-b2f7-c40aca1f87e0